porting resnet
